Update database.py
https://github.com/sininspira2/spice-tracker-bot-tma/pull/5#discussion_r2323721616

There is a potential race condition in the guild_withdraw function. The balance check is performed before the database transaction starts. If two guild_withdraw commands run concurrently, both might pass the balance check before either one updates the treasury, which could lead to a negative balance. This is a classic Time-of-check to time-of-use (TOCTOU) vulnerability.

To make the operation atomic, you should perform the check and the update within the same transaction. A robust way to do this is to modify the UPDATE statement to fail if funds are insufficient and then check the number of affected rows.